Fix various issues of the last release
